feat: 내 댓글 삭제기능 구현 [ISSUE-469] (#482)
* feat: 댓글 삭제버튼 위치 레이아웃 잡음

* refactor: 댓글 작성자 이름 글자 크기 축소

* feat: 댓글 삭제 api 구현

* feat: 댓글 삭제 버튼과 댓글 삭제 api 연결

* refactor: 빙봉 피드백 반영 및 글삭제 confirm 창, 댓글삭제 성공 alert 창 추가

+++ front-end/legeno-around-here/src/components/pages/post/DeletePostButton.js
@@ -3,7 +3,9 @@ import Button from '@material-ui/core/Button';
 import { deletePost } from '../../api/API';
 
 const onDeletePost = (accessToken, postId, history) => {
-  deletePost(accessToken, postId, history);
+  if (window.confirm("정말 삭제하시겠습니까?")) {
+    deletePost(accessToken, postId, history)
+  }
 };
 
 const DeletePostButton = ({ accessToken, postId, history }) => {
